updated docs and unit tests
